+++ src/collective/collectionfilter/portlets/collectionsearch.py
@@ -1,10 +1,9 @@
 # -*- coding: utf-8 -*-
-from Products.CMFPlone.utils import get_top_request
-from Products.CMFPlone.utils import safe_unicode
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from collective.collectionfilter import _
 from collective.collectionfilter.baseviews import BaseSearchView
 from collective.collectionfilter.interfaces import ICollectionSearchSchema
+from collective.collectionfilter.portlets import BasePortletRenderer
 from plone.app.portlets.portlets import base
 from plone.portlets.interfaces import IPortletDataProvider
 from zope.interface import implementer
@@ -43,25 +42,8 @@ class Assignment(base.Assignment):
             return _(u'Collection Search')
 
 
-class Renderer(base.Renderer, BaseSearchView):
+class Renderer(BasePortletRenderer, BaseSearchView):
     render = ViewPageTemplateFile('collectionsearch.pt')
-
-    @property
-    def filter_id(self):
-        request = get_top_request(self.request)
-        portlethash = request.form.get(
-            'portlethash',
-            getattr(self, '__portlet_metadata__', {}).get('hash', '')
-        )
-        return portlethash
-
-    @property
-    def reload_url(self):
-        reload_url = '{0}/@@render-portlet?portlethash={1}'.format(
-            self.context.absolute_url(),
-            safe_unicode(self.filter_id),
-        )
-        return reload_url
 
 
 class AddForm(base.AddForm):