+++ src/collective/collectionfilter/utils.py
@@ -28,7 +28,7 @@ def safe_decode(val):
     """
     ret = val
     if isinstance(val, dict):
-        ret = dict([(safe_decode(k), safe_decode(v)) for k, v in val.items()])
+        ret = dict([(safe_decode(k), safe_decode(v)) for k, v in val.items() if v])  # noqa
     elif isinstance(val, list):
         ret = [safe_decode(it) for it in val]
     elif isinstance(val, tuple):
@@ -43,7 +43,7 @@ def safe_encode(val):
     """
     ret = val
     if isinstance(val, dict):
-        ret = dict([(safe_encode(k), safe_encode(v)) for k, v in val.items()])
+        ret = dict([(safe_encode(k), safe_encode(v)) for k, v in val.items() if v])  # noqa
     elif isinstance(val, list):
         ret = [safe_encode(it) for it in val]
     elif isinstance(val, tuple):
@@ -82,7 +82,7 @@ def base_query(request_params={}, extra_ignores=[]):
         'batch',
         'sort_on',
         'limit',
-        'portlethash'
+        'portlethash',
     ] + extra_ignores
     # Now remove all to-be-ignored request parameters.
     urlquery = {